Always use router to set from/to params
This prevent the route to be pushed twice in history.
